fix(estring): fix comparison issue in boyer moore search algorithm

+++ src/estring.c
@@ -11,10 +11,10 @@ int *bad_char_table(const char *F) {
     return NULL;
 
   int *table = (int *)malloc(256 * sizeof(int)); // 256 for everu ASCII char
-  for (int i = 0; i < 256; i++)
+  for (size_t i = 0; i < 256; i++)
     table[i] = -1;
 
-  for (int i = 0; i < strlen(F); i++)
+  for (size_t i = 0; i < strlen(F); i++)
     table[(unsigned char)F[i]] = i;
 
   return table;
@@ -24,12 +24,12 @@ int boyer_moore_search(const char *T, const char *F) {
   if (!T || !F)
     return NULL_POINTER;
 
-  int n = strlen(T);
-  int m = strlen(F);
+  size_t n = strlen(T);
+  size_t m = strlen(F);
   int pos = -1;
 
   int *bad_char = bad_char_table(F);
-  int s = 0;
+  size_t s = 0;
   while (s <= n - m) {
     int j = m - 1;
     while (j >= 0 && F[j] == T[s + j])
